FIX CALENDER

+++ app/models/ticket.php
@@ -6,7 +6,7 @@ class Ticket implements \JsonSerializable
     private int $id;
     private string $title;
     private string $type;
-    private int $price;
+    private float $price;
     private Event $event;
 
     #[ReturnTypeWillChange]
@@ -43,11 +43,11 @@ class Ticket implements \JsonSerializable
         $this->type = $type;
         return $this;
     }
-    public function getPrice(): int
+    public function getPrice(): float
     {
         return $this->price;
     }
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
         return $this;
